Confidence Interval - Bias-Corrected and Accelerated

Calculates bias-corrected and accelerated confidence intervals.

bca(
  thetahatstar,
  thetahat,
  data,
  fitFUN,
  alpha = c(0.001, 0.01, 0.05),
  wald = FALSE,
  null = 0,
  dist = "z",
  df,
  eval = FALSE,
  theta = 0,
  par = FALSE,
  ncores = NULL,
  ...
)

Arguments

thetahatstar
Numeric vector. The bootstrap sampling distribution \(\left( \boldsymbol{\hat{\theta}^{*}} \right)\), that is, the sampling distribution of thetahat estimated for each b bootstrap sample. \( \hat{\theta}_{\left( 1 \right)}, \hat{\theta}_{\left( 2 \right)}, \hat{\theta}_{\left( 3 \right)}, \hat{\theta}_{\left( b \right)}, \dots, \hat{\theta}_{\left( B \right)} \) .

thetahat
Numeric. Parameter estimate \(\left( \hat{\theta} \right)\) from the original sample data.

data
Vector, matrix, or data frame. Sample data. Ignored if thetahatstarjack is provided.

fitFUN
Function. Fit function to use on data. The first argument should correspond to data. Other arguments can be passed to fitFUN using .... fitFUN should return a single value. Ignored if thetahatstarjack is provided.

alpha
Numeric vector. Significance level \(\left( \alpha \right)\) . By default, alpha is set to conventional significance levels alpha = c(0.001, 0.01, 0.05).

wald
Logical. If TRUE, calculates the square root of the Wald test statistic and p-value. The estimated bootstrap standard error is used. The arguments null, dist, and df are used to calculate the square root of the Wald test statistic and p-value and are NOT used in constructing the bootstrap confidence interval. If FALSE, returns statistic = NA and p = NA If FALSE, the arguments null, dist, and df are ignored.

null
Numeric. Hypothesized value of theta \(\left( \theta_{0} \right)\). Set to zero by default.

dist
Character string. dist = "z" for the standard normal distribution. dist = "t" for the t distribution.

df
Numeric. Degrees of freedom (df) if dist = "t". Ignored if dist = "z".

eval
Logical. Evaluate confidence intervals using zero_hit(), theta_hit(), len(), and shape().

theta
Numeric. Population parameter \(\left( \theta \right)\).

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

...
Arguments to pass to fitFUN.

Value

Returns a vector with the following elements:

statistic

Square root of Wald test statistic. NA if wald = FALSE.

p

p-value. NA if wald = FALSE.

se

Estimated bootstrap standard error \(\left( \widehat{\mathrm{se}}_{\mathrm{B}} \left( \hat{\theta} \right) \right)\).

ci_

Estimated bias-corrected and accelerated confidence limits corresponding to alpha from the bootstrap sampling distribution thetahatstar \(\left( \boldsymbol{\hat{\theta}^{*}} \right)\).

If eval = TRUE, appends the following to the results vector

zero_hit_

Logical. Tests if confidence interval contains zero.

theta_hit_

Logical. Tests if confidence interval contains theta.

length_

Length of confidence interval.

shape_

Shape of confidence interval.

Details

The estimated bootstrap standard error is given by $$ \widehat{\mathrm{se}}_{\mathrm{B}} \left( \hat{\theta} \right) = \sqrt{ \frac{1}{B - 1} \sum_{b = 1}^{B} \left[ \hat{\theta}^{*} \left( b \right) - \hat{\theta}^{*} \left( \cdot \right) \right]^2 } $$ where $$ \hat{\theta}^{*} \left( \cdot \right) = \frac{1}{B} \sum_{b = 1}^{B} \hat{\theta}^{*} \left( b \right) . $$

In addition to the bias-correction \(\hat{z}_{0}\) discussed in bc(), the acceleration \(\hat{a}\), which refers to the rate of change of the standard error of \(\hat{\theta}\) with respect to the true parameter value \(\theta\), is factored in.

The acceleration \(\hat{a}\) is given by

$$ \hat{a} = \frac{ \sum_{i = 1}^{n} \left[ \hat{\theta}_{ \left( \cdot \right) } - \hat{\theta}_{ \left( i \right) } \right]^{3} } { 6 \left\{ \sum_{i = 1}^{n} \left[ \hat{\theta}_{ \left( \cdot \right) } - \hat{\theta}_{ \left( i \right)} \right]^{2} \right\}^{3/2} } $$ where $$ \hat{\theta}_{ \left( i \right) } = \left\{ \hat{\theta}_{\left( 1 \right)}, \hat{\theta}_{\left( 2 \right)}, \hat{\theta}_{\left( 3 \right)}, \dots, \hat{\theta}_{\left( n \right)} \right\} $$ is the jackknife sampling distribution and $$ \hat{\theta}_{ \left( \cdot \right) } = \frac{1}{n} \sum_{i = 1}^{n} \hat{\theta}_{ \left( i \right) } $$ is the jackknife mean. See jack() and jack_hat() .

Using \(\hat{z}_{0}\) and \(\hat{a}\) we can obtain the adjusted \(z\)-scores as follows

$$ z_{ \mathrm{BCa}_{\mathrm{lo}} } = \hat{z}_{0} + \frac{ \hat{z}_{0} + z_{ \left( \frac{ \alpha } { 2 } \right) } } { 1 - \hat{a} \left[ \hat{z}_{0} + z_{ \left( \frac{ \alpha } { 2 } \right) } \right] } , $$

$$ z_{ \mathrm{BCa}_{\mathrm{up}} } = \hat{z}_{0} + \frac{ \hat{z}_{0} + z_{ \left[ 1 - \frac{ \alpha } { 2 } \right] } } { 1 - \hat{a} \left[ \hat{z}_{0} + z_{ \left( 1 - \frac{ \alpha } { 2 } \right) } \right] } . $$

The adjusted \(z\)-scores are used to determine the adjusted percentile ranks to form the confidence interval.

The bias-corrected and accelerated confidence interval is given by $$ \left[ \hat{\theta}_{\mathrm{lo}}, \hat{\theta}_{\mathrm{up}} \right] = \left[ \hat{\theta}^{*}_{ \left( z_{ \mathrm{BCa}_{ \mathrm{lo} } } \right) }, \hat{\theta}^{*}_{ \left( z_{ \mathrm{BCa}_{ \mathrm{up} } } \right) } \right] . $$

For more details and examples see the following vignettes:

Notes: Introduction to Nonparametric Bootstrapping

Notes: Introduction to Parametric Bootstrapping

References

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

See also