Confidence Intervals to Critical Values

Calculates the \(z\), \(t\), \(\chi^2\), or \(F\) critical value/s of confidence limits associated with the specified confidence interval/s.

ci2crit(
  ci = c(0.999, 0.99, 0.95),
  dist = "z",
  two.tailed = TRUE,
  right.tail = TRUE,
  ...
)

Arguments

ci
Numeric vector. Confidence interval. By default, ci is set to conventional confidence intervals ci = c(0.999, 0.99, 0.95).

dist
Character string. dist = "z" for the standard normal distribution. dist = "t" for the t distribution. dist = "F" for the F distribution. dist = "chisq" for the chi-square distribution.

two.tailed
Logical. If TRUE, two-tailed alpha. If FALSE, one-tailed alpha. Ignored if dist = "F" or dist = "chisq" as both tests are one-tailed using the right tail.

right.tail
Logical. If TRUE, right tail (positive critical value). If FALSE, left tail (negative critical value). Ignored if two.tailed = TRUE. Ignored if dist = "F" or dist = "chisq" as both tests are one-tailed using the right tail.

...
Degrees of freedom. df for dist = "t" and dist = "chisq". df1 and df2 for dist = "F".

Value

Returns \(z\), \(t\), \(\chi^2\), or \(F\) critical value/s associated with the specified confidence interval/s. The results are sorted from smallest to largest.

References

Wikipedia: Statistical significance

Wikipedia: Confidence interval

See also

Examples

# vector
## two-tailed
### z
ci2crit(
  ci = 0.95,
  dist = "z"
)

#> [1] -1.959964  1.959964

### t
ci2crit(
  ci = 0.95,
  dist = "t",
  df = 5
)

#> [1] -2.570582  2.570582

## one-tailed
### right
#### z
ci2crit(
  ci = 0.95,
  dist = "z",
  two.tailed = FALSE
)

#> [1] 1.644854

#### t
ci2crit(
  ci = 0.95,
  dist = "t",
  two.tailed = FALSE,
  df = 5
)

#> [1] 2.015048

#### F
ci2crit(
  ci = 0.95,
  dist = "F",
  two.tailed = FALSE,
  df1 = 5,
  df2 = 5
)

#> [1] 5.050329

#### chi-square
ci2crit(
  ci = 0.95,
  dist = "chisq",
  two.tailed = FALSE,
  df = 5
)

#> [1] 11.0705

### left
#### z
ci2crit(
  ci = 0.95,
  dist = "z",
  two.tailed = FALSE,
  right.tail = FALSE
)

#> [1] -1.644854

#### t
ci2crit(
  ci = 0.95,
  dist = "t",
  two.tailed = FALSE,
  right.tail = FALSE,
  df = 5
)

#> [1] -2.015048