Confidence Intervals to Probabilities

Calculates the cumulative probabilities of confidence limits associated with the specified confidence interval/s for a two-tailed test.

ci2prob(ci = c(0.999, 0.99, 0.95))

Arguments

ci
Numeric vector. Confidence interval. By default, ci is set to conventional confidence intervals ci = c(0.999, 0.99, 0.95).

Value

Returns probabilities associated with the specified confidence interval/s. The results are sorted from smallest to largest.

References

Wikipedia: Statistical significance

Wikipedia: Confidence interval

See also

Examples

# vector
ci2prob(ci = c(0.999, 0.99, 0.95))

#> [1] 0.0005 0.0050 0.0250 0.9750 0.9950 0.9995

# single numeric value
ci2prob(ci = 0.95)

#> [1] 0.025 0.975