Parametric Bootstrap (Multivariate Normal) from \(\boldsymbol{\hat{\Sigma}}\) and \(\boldsymbol{\hat{\mu}}\)

Generates B number of parametric bootstrap samples from the original sample data. data is referred to as the empirical distribution \( \hat{F}_{ \mathrm{MVN} \left( \boldsymbol{\hat{\Sigma}}, \boldsymbol{\hat{\mu}} \right) } \) . Data is generated from a multivariate normal distribution using the estimated variance-covariance matrix \(\boldsymbol{\hat{\Sigma}}\) and mean vector \(\boldsymbol{\hat{\mu}}\) .

.pb_mvn(n, Sigmahat, muhat, B = 2000L, par = FALSE, ncores = NULL)

Arguments

n
Integer. Sample size.

Sigmahat
Matrix. Estimated variance-covariance matrix from the original sample data.

muhat
Vector. Estimated mean vector from the original sample data.

B
Integer. Number of bootstrap samples.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

Value

Returns a list of length B of parametric bootstrap samples.

Details

For more details and examples see the following vignettes:

Notes: Introduction to Nonparametric Bootstrapping

Notes: Introduction to Parametric Bootstrapping

References

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

Wikipedia: Bootstrapping (statistics)

See also

Examples

B <- 5L
Sigmahat <- matrix(
  data = c(
    82.37344,
    70.55922,
    17.83930,
    70.55922,
    112.57145,
    -75.98558,
    17.83930,
    -75.98558,
    338.46263
  ),
  nrow = 3
)
muhat <- c(
  108.3060,
  105.3324,
  103.4009
)
Xstar <- .pb_mvn(
  n = 5,
  Sigmahat = Sigmahat,
  muhat = muhat,
  B = B
)
str(Xstar)

#> List of 5
#>  $ : num [1:5, 1:3] 117.2 94.1 112.2 101.9 107 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:5, 1:3] 112.6 109.1 96.9 112.4 121 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:5, 1:3] 107 90 106 105 119 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:5, 1:3] 117 100 113 107 110 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ : num [1:5, 1:3] 95.7 118.5 112.7 113.2 97.3 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL