Nonparametric Bootstrap

Generates B number of nonparametric bootstrap samples from the original sample data. data is referred to as the empirical distribution \(\hat{F}\).

nb(data, B = 2000L, par = FALSE, ncores = NULL)

Arguments

data
Vector, matrix or data frame. Sample data to bootstrap. The empirical distribution \(\hat{F}\).

B
Integer. Number of bootstrap samples.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

Value

Returns a list of length B of nonparametric bootstrap samples.

Details

For more details and examples see the following vignettes:

Notes: Introduction to Nonparametric Bootstrapping

Notes: Introduction to Parametric Bootstrapping

References

Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

Wikipedia: Bootstrapping (statistics)

See also

Examples

B <- 5L
n <- 5
#################################
# vector
#################################
x <- rnorm(n = n)
x_star <- nb(
  data = x,
  B = B
)
str(x_star)

#> List of 5
#>  $ : num [1:5] 0.693 1.031 0.335 1.031 0.693
#>  $ : num [1:5] 0.693 0.693 0.335 0.335 1.345
#>  $ : num [1:5] 1.031 0.335 1.345 0.693 1.345
#>  $ : num [1:5] 0.693 0.693 1.031 2.631 2.631
#>  $ : num [1:5] 1.031 0.335 0.693 0.693 1.345

#################################
# matrix
#################################
x1 <- rnorm(n = n)
x2 <- rnorm(n = n)
x3 <- rnorm(n = n)
X <- cbind(x1, x2, x3)
X_star <- nb(
  data = X,
  B = B
)
str(X_star)

#> List of 5
#>  $ : num [1:5, 1:3] 1.604 1.604 1.189 0.212 0.207 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:5, 1:3] 0.207 1.189 0.212 0.473 0.207 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:5, 1:3] 0.212 1.604 0.473 1.604 0.473 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:5, 1:3] 1.189 0.473 0.212 1.189 1.604 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:5, 1:3] 0.207 1.604 1.189 0.473 0.207 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"

#################################
# data frame
#################################
X <- as.data.frame(X)
X_star <- nb(
  data = X,
  B = B
)
str(X_star)

#> List of 5
#>  $ :'data.frame':	5 obs. of  3 variables:
#>   ..$ x1: num [1:5] 1.189 1.604 1.189 1.189 0.473
#>   ..$ x2: num [1:5] -0.918 0.293 -0.918 -0.918 -0.302
#>   ..$ x3: num [1:5] 0.251 -0.662 0.251 0.251 0.199
#>  $ :'data.frame':	5 obs. of  3 variables:
#>   ..$ x1: num [1:5] 1.189 0.207 0.207 0.207 0.212
#>   ..$ x2: num [1:5] -0.9179 -0.2871 -0.2871 -0.2871 -0.0444
#>   ..$ x3: num [1:5] 0.251 -0.518 -0.518 -0.518 -1.269
#>  $ :'data.frame':	5 obs. of  3 variables:
#>   ..$ x1: num [1:5] 1.189 0.473 1.189 1.189 1.189
#>   ..$ x2: num [1:5] -0.918 -0.302 -0.918 -0.918 -0.918
#>   ..$ x3: num [1:5] 0.251 0.199 0.251 0.251 0.251
#>  $ :'data.frame':	5 obs. of  3 variables:
#>   ..$ x1: num [1:5] 0.207 0.473 0.473 0.473 0.212
#>   ..$ x2: num [1:5] -0.2871 -0.3024 -0.3024 -0.3024 -0.0444
#>   ..$ x3: num [1:5] -0.518 0.199 0.199 0.199 -1.269
#>  $ :'data.frame':	5 obs. of  3 variables:
#>   ..$ x1: num [1:5] 0.212 0.473 1.604 0.473 0.212
#>   ..$ x2: num [1:5] -0.0444 -0.3024 0.2926 -0.3024 -0.0444
#>   ..$ x3: num [1:5] -1.269 0.199 -0.662 0.199 -1.269