Jackknife

Generates jackknife samples.

jack(data, par = FALSE, ncores = NULL)

Arguments

data
Vector, matrix, or data frame. Sample data.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

Value

Returns a list of jackknife samples of length n, where n is the sample size of data. Each item in the list, will have a sample size of n - 1.

Details

$$ \mathbf{x}_i = \left\{ x_1, x_2, x_3, \dots, x_n \right\} $$ for \(i = \left\{ 1, 2, 3, \dots, n \right\}\) the \(i\)th jackknife sample consists of the original sample data with the \(i\)th observation removed.

References

Wikipedia: Jackknife resampling

See also

Examples

n <- 5
#################################
# vector
#################################
x <- rnorm(n = n)
x_star <- jack(
  data = x
)
str(x_star)

#> List of 5
#>  $ : num [1:4] 0.6067 -0.1099 0.1722 -0.0903
#>  $ : num [1:4] 0.5239 -0.1099 0.1722 -0.0903
#>  $ : num [1:4] 0.5239 0.6067 0.1722 -0.0903
#>  $ : num [1:4] 0.5239 0.6067 -0.1099 -0.0903
#>  $ : num [1:4] 0.524 0.607 -0.11 0.172

#################################
# matrix
#################################
x1 <- rnorm(n = n)
x2 <- rnorm(n = n)
x3 <- rnorm(n = n)
X <- cbind(x1, x2, x3)
X_star <- jack(
  data = X
)
str(X_star)

#> List of 5
#>  $ : num [1:4, 1:3] 1.298 0.749 0.556 -0.548 -2.612 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:4, 1:3] 1.924 0.749 0.556 -0.548 1.111 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:4, 1:3] 1.924 1.298 0.556 -0.548 1.111 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:4, 1:3] 1.924 1.298 0.749 -0.548 1.111 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"
#>  $ : num [1:4, 1:3] 1.924 1.298 0.749 0.556 1.111 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:3] "x1" "x2" "x3"

#################################
# data frame
#################################
X <- as.data.frame(X)
X_star <- jack(
  data = X
)
str(X_star)

#> List of 5
#>  $ :'data.frame':	4 obs. of  3 variables:
#>   ..$ x1: num [1:4] 1.298 0.749 0.556 -0.548
#>   ..$ x2: num [1:4] -2.612 -0.156 0.434 -0.382
#>   ..$ x3: num [1:4] 1.0631 1.0487 -0.0381 0.4861
#>  $ :'data.frame':	4 obs. of  3 variables:
#>   ..$ x1: num [1:4] 1.924 0.749 0.556 -0.548
#>   ..$ x2: num [1:4] 1.111 -0.156 0.434 -0.382
#>   ..$ x3: num [1:4] 0.4242 1.0487 -0.0381 0.4861
#>  $ :'data.frame':	4 obs. of  3 variables:
#>   ..$ x1: num [1:4] 1.924 1.298 0.556 -0.548
#>   ..$ x2: num [1:4] 1.111 -2.612 0.434 -0.382
#>   ..$ x3: num [1:4] 0.4242 1.0631 -0.0381 0.4861
#>  $ :'data.frame':	4 obs. of  3 variables:
#>   ..$ x1: num [1:4] 1.924 1.298 0.749 -0.548
#>   ..$ x2: num [1:4] 1.111 -2.612 -0.156 -0.382
#>   ..$ x3: num [1:4] 0.424 1.063 1.049 0.486
#>  $ :'data.frame':	4 obs. of  3 variables:
#>   ..$ x1: num [1:4] 1.924 1.298 0.749 0.556
#>   ..$ x2: num [1:4] 1.111 -2.612 -0.156 0.434
#>   ..$ x3: num [1:4] 0.4242 1.0631 1.0487 -0.0381